perf(ci): split 3 heaviest test files + optimise shard seed
