+++ package.json
@@ -67,14 +67,14 @@
     "test:playwright": "npx playwright test",
     "journey": "node tests/journeys/_runner.mjs",
     "verify": "npm test && npm run check && npm run capacity:verify-evidence",
-    "verify:grammar-qg": "npm run audit:grammar-qg && npm run audit:grammar-qg:deep && node --test tests/grammar-question-generator-audit.test.js tests/grammar-qg-p5-depth.test.js tests/grammar-qg-p5-content-quality.test.js tests/grammar-qg-p5-report-validation.test.js tests/grammar-functionality-completeness.test.js tests/grammar-production-smoke.test.js tests/grammar-selection.test.js tests/grammar-engine.test.js",
-    "verify:grammar-qg-p6": "npm run audit:grammar-qg && npm run audit:grammar-qg:deep && node --test tests/grammar-question-generator-audit.test.js tests/grammar-qg-p5-depth.test.js tests/grammar-qg-p5-content-quality.test.js tests/grammar-qg-p5-report-validation.test.js tests/grammar-qg-p6-governance.test.js tests/grammar-qg-p6-telemetry.test.js tests/grammar-qg-p6-health-report.test.js tests/grammar-qg-p6-mixed-transfer-calibration.test.js tests/grammar-qg-p6-retention.test.js tests/grammar-qg-p6-review-pack.test.js tests/grammar-functionality-completeness.test.js tests/grammar-production-smoke.test.js tests/grammar-selection.test.js tests/grammar-engine.test.js",
+    "verify:grammar-qg": "npm run audit:grammar-qg && npm run audit:grammar-qg:deep && node --test tests/grammar-question-generator-audit.test.js tests/grammar-qg-p5-depth.test.js tests/grammar-qg-p5-content-quality.test.js tests/grammar-qg-p5-report-validation.test.js tests/grammar-functionality-completeness.test.js tests/grammar-production-smoke.test.js tests/grammar-selection-core.test.js tests/grammar-selection-edge.test.js tests/grammar-engine.test.js",
+    "verify:grammar-qg-p6": "npm run audit:grammar-qg && npm run audit:grammar-qg:deep && node --test tests/grammar-question-generator-audit.test.js tests/grammar-qg-p5-depth.test.js tests/grammar-qg-p5-content-quality.test.js tests/grammar-qg-p5-report-validation.test.js tests/grammar-qg-p6-governance.test.js tests/grammar-qg-p6-telemetry.test.js tests/grammar-qg-p6-health-report.test.js tests/grammar-qg-p6-mixed-transfer-calibration.test.js tests/grammar-qg-p6-retention.test.js tests/grammar-qg-p6-review-pack.test.js tests/grammar-functionality-completeness.test.js tests/grammar-production-smoke.test.js tests/grammar-selection-core.test.js tests/grammar-selection-edge.test.js tests/grammar-engine.test.js",
     "verify:grammar-qg-p7": "npm run verify:grammar-qg-p6 && node --test tests/grammar-qg-p7-governance.test.js tests/grammar-qg-p7-elapsed-timing.test.js tests/grammar-qg-p7-event-expansion.test.js tests/grammar-qg-p7-health-report.test.js tests/grammar-qg-p7-action-candidates.test.js tests/grammar-qg-p7-production-evidence.test.js tests/grammar-qg-p7-analytics-view.test.js",
     "verify:grammar-qg-p8": "npm run verify:grammar-qg-p7 && node --test tests/grammar-qg-p8-question-quality.test.js tests/grammar-qg-p8-governance.test.js tests/grammar-qg-p8-oracles.test.js tests/grammar-qg-p8-review-register.test.js tests/grammar-qg-p8-ux-support.test.js",
     "verify:grammar-qg-p9": "npm run verify:grammar-qg-p8 && node --test tests/grammar-qg-p9-report-evidence-lock.test.js tests/grammar-qg-p9-inventory-manifest.test.js tests/grammar-qg-p9-review-evidence.test.js tests/grammar-qg-p9-learner-surface.test.js tests/grammar-qg-p9-table-choice-contract.test.js tests/grammar-qg-p9-event-expansion-real-tags.test.js tests/grammar-qg-p9-oracle-windows.test.js tests/grammar-qg-p9-blocklist-scheduler.test.js",
     "verify:grammar-qg-p10": "npm run verify:grammar-qg-p9 && node --test tests/grammar-qg-p10-evidence-truth.test.js tests/grammar-qg-p10-prompt-cue-contract.test.js tests/grammar-qg-p10-read-aloud-alignment.test.js tests/grammar-qg-p10-table-render.test.js tests/grammar-qg-p10-render-surface.test.js tests/grammar-qg-p10-scheduler-safety.test.js tests/grammar-qg-p10-evidence-artefacts.test.js",
     "verify:grammar-qg-p11": "npm run verify:grammar-qg-p10 && node --test tests/grammar-qg-p11-scheduler-blocklist.test.js tests/grammar-qg-p11-production-smoke-contract.test.js tests/grammar-qg-p11-verify-chain.test.js tests/grammar-qg-p11-evidence-truth.test.js tests/grammar-qg-p11-distractor-matrix-truth.test.js tests/grammar-qg-p11-target-extraction.test.js tests/grammar-qg-p11-accessibility-copy.test.js",
-    "verify:grammar-qg-p14": "npm run verify:grammar-qg-p11 && node --test tests/grammar-qg-p13-runtime-certification.test.js tests/grammar-qg-p14-depth-variety.test.js tests/grammar-qg-p14-surface-telemetry.test.js && npm run audit:grammar-qg:semantic",
+    "verify:grammar-qg-p14": "npm run verify:grammar-qg-p11 && node --test tests/grammar-qg-p13-runtime-certification.test.js tests/grammar-qg-p14-depth-variety.test.js tests/grammar-qg-p14-surface-telemetry-a.test.js tests/grammar-qg-p14-surface-telemetry-b.test.js tests/grammar-qg-p14-surface-telemetry-c.test.js && npm run audit:grammar-qg:semantic",
     "verify:grammar-qg-production-release": "npm run verify:grammar-qg-p14 && node scripts/generate-grammar-manual-expansion.mjs --check && node scripts/validate-grammar-qg-certification-evidence.mjs reports/grammar/grammar-qg-p19-certification-manifest.json --expected-release=grammar-qg-p19-2026-05-04 && npm run audit:grammar-qg:open-response-fairness && npm run audit:grammar-qg:smart-practice",
     "verify:grammar-qg-p10-production-release-historical": "node scripts/validate-grammar-qg-certification-evidence.mjs reports/grammar/grammar-qg-p10-certification-manifest.json --expected-release=grammar-qg-p10-2026-04-29",
     "grammar:qg:calibrate": "node scripts/grammar-qg-calibrate.mjs",
