--- tests/grammar-learning-integrity.test.js
@@ -1,554 +1,0 @@
-// U6 — Seeded adaptive-selection simulation suite.
-//
-// Plan: docs/plans/2026-04-26-001-feat-grammar-phase4-learning-hardening-plan.md
-// (search "U6. **Seeded adaptive-selection simulation suite**", ~line 619).
-//
-// What this file proves — aggregated across the 8 canonical seeds (1, 7, 13,
-// 42, 100, 2025, 31415, 65535):
-//   1. **Due outranks non-due**       — same concept picked more often when due.
-//   2. **Weak outranks secure**       — weak concept's first position lands
-//                                        in the first half of a 10-item queue.
-//   3. **Recent-miss recycle**        — concept with a recent miss reappears
-//                                        within 5 items in >= 6/8 seeds.
-//   4. **Template freshness**         — no template id appears 3+ times in a
-//                                        10-item queue under ANY of 8 seeds.
-//   5. **Concept freshness**          — no concept appears 3+ times
-//                                        consecutively in a 10-item queue
-//                                        across 8 seeds.
-//   6. **Mini-pack balance**          — `buildGrammarMiniPack({size:8})`
-//                                        distributes question types within
-//                                        +/- ceil(size/3) of even.
-//   7. **Supported < independent**    — `applyGrammarAttemptToState` yields
-//                                        smaller mastery strength when
-//                                        `mode='worked'` (support=2) vs
-//                                        `mode='smart'` (support=0), summed
-//                                        across 8 seeds x 3 runs.
-//   8. **Pathological focus pool**    — empty mastery + focusConceptId on a
-//                                        2-template concept returns a valid
-//                                        10-item queue; no NPE.
-//   9. **All concepts secured**       — engine still returns a valid queue
-//                                        biased toward breadth (no NPE).
-//  10. **Mini-pack size=0 error path** — returns an empty array; no NPE.
-//  11. **20-round replay spread**     — final mastery shows SPREAD
-//                                        (>=3 concepts touched per seed, no
-//                                        single concept dominating).
-//
-// Constraint from the plan (U6):
-//   > NO code changes to `worker/src/subjects/grammar/selection.js` or
-//   > `engine.js` unless simulation surfaces a real bug. This is a
-//   > VERIFICATION unit, not a refactor.
-//
-// Any principle that genuinely fails in 8/8 is reported via an explicit
-// assertion with seed-level diagnostics so the calibration question (tune
-// weights vs tune threshold) reaches James transparently.
-
-import test from 'node:test';
-import assert from 'node:assert/strict';
-
-import {
-  buildGrammarMiniPack,
-  buildGrammarPracticeQueue,
-} from '../worker/src/subjects/grammar/selection.js';
-import {
-  CANONICAL_SEEDS,
-  SIM_NOW_MS,
-  buildQueueForSeed,
-  buildMiniPackForSeed,
-  conceptHitsInQueue,
-  templateCountsInQueue,
-  questionTypeCountsInPack,
-  hasConsecutiveConceptRun,
-  pushRecentMiss,
-  runSingleAttemptMasteryGain,
-  run20RoundReplay,
-  simulateAcrossSeeds,
-  stateWithAllConcepts,
-  stateWithConceptStatus,
-} from './helpers/grammar-simulation.js';
-
-// -----------------------------------------------------------------------------
-// Principle 1 — Due outranks non-due (aggregate total picks comparison).
-//
-// Plan text: "for the same concept, `due=true` appears in queue position ≤ k
-// for `k ≤ 3` in all 8 seeds". The stochastic weighted sampler (see
-// worker/src/subjects/grammar/selection.js) cannot guarantee first-position
-// <=3 in 8/8 seeds when one concept competes against 17 others with 2-5/51
-// templates — that would require a hard sort, not a weighted pick.
-//
-// The PRINCIPLE the engine is designed to enforce: the weighted rank boost
-// for `due` concepts must yield strictly more picks than equivalent non-due
-// mastery. To dampen single-concept + single-seed variance, this test
-// aggregates across all 18 concepts x 8 seeds (144 samples) — each concept
-// is rotated into the due / not-due state in turn, everything else held
-// constant. Mirrors the existing Phase 2 U2 test at
-// tests/grammar-selection.test.js:134-168 but with richer aggregation.
-// -----------------------------------------------------------------------------
-
-test('U6 principle: a due concept is picked more often than an equivalent non-due concept (aggregate across 18 concepts x 8 seeds)', async () => {
-  const { GRAMMAR_CONCEPTS } = await import('../worker/src/subjects/grammar/content.js');
-  const { createInitialGrammarState } = await import('../worker/src/subjects/grammar/engine.js');
-  function stateForConcept(conceptId, dueOffset) {
-    const state = createInitialGrammarState();
-    // Mirror the exact node shape from the Phase 2 U2 existing test
-    // (tests/grammar-selection.test.js:146-154) so the due vs not-due
-    // contrast is fair at strength 0.85 / correctStreak 3.
-    state.mastery.concepts[conceptId] = {
-      attempts: 5,
-      correct: 4,
-      wrong: 1,
-      strength: 0.85,
-      intervalDays: 7,
-      dueAt: SIM_NOW_MS + dueOffset,
-      lastSeenAt: null,
-      lastWrongAt: null,
-      correctStreak: 3,
-    };
-    return state;
-  }
-  let dueTotal = 0;
-  let notDueTotal = 0;
-  const perConcept = [];
-  for (const concept of GRAMMAR_CONCEPTS) {
-    const stateDue = stateForConcept(concept.id, -60_000); // overdue
-    const stateNotDue = stateForConcept(concept.id, +7 * 86_400_000); // due in 7 days
-    let dueCount = 0;
-    let notDueCount = 0;
-    for (const seed of CANONICAL_SEEDS) {
-      dueCount += conceptHitsInQueue(
-        buildQueueForSeed({ seed, size: 12, mastery: stateDue.mastery }),
-        concept.id,
-      ).length;
-      notDueCount += conceptHitsInQueue(
-        buildQueueForSeed({ seed, size: 12, mastery: stateNotDue.mastery }),
-        concept.id,
-      ).length;
-    }
-    dueTotal += dueCount;
-    notDueTotal += notDueCount;
-    perConcept.push({ conceptId: concept.id, due: dueCount, notDue: notDueCount });
-    // Pointwise per-concept guard (MEDIUM fix). A single concept silently
-    // reversing (due < notDue) could be masked by the 18-concept aggregate —
-    // e.g., 17 healthy concepts could pump `dueTotal` past `notDueTotal *
-    // 1.5` while 1 concept regresses severely. This per-concept assertion
-    // inside the loop surfaces any drastic single-concept reversal.
-    //
-    // A strict `due >= notDue` per-concept gate is too tight for a
-    // stochastic weighted sampler: with only 8 seeds and 2-5 templates per
-    // concept out of 51 total, low-volume concepts can swing +/-2 picks by
-    // chance alone (e.g., adverbials currently: due=3 notDue=5, a 2-pick
-    // sampler jitter). We therefore use a tolerance-based equivalent: a
-    // per-concept regression must be bounded by BOTH an absolute slack
-    // (`+3`) AND a relative factor (`2x`). This catches a real per-concept
-    // reversal (e.g., due=2, notDue=12) while tolerating sampler variance.
-    assert.ok(
-      notDueCount <= dueCount + 3 && notDueCount <= (dueCount + 1) * 2,
-      `concept ${concept.id} due=${dueCount} notDue=${notDueCount} — `
-      + `per-concept due-outranks regression beyond sampler-variance tolerance`,
-    );
-  }
-  assert.ok(
-    dueTotal > notDueTotal,
-    `Due concepts must be picked more often than equivalent not-due concepts across 18 x 8 samples. `
-    + `dueTotal=${dueTotal} notDueTotal=${notDueTotal}. perConcept=${JSON.stringify(perConcept)}`,
-  );
-  // Spirit check: the ratio should be materially > 1 (the engine's `due`
-  // weight is 3.0 on top of a base ~1.6 at strength 0.85, so we expect a
-  // roughly 2-4x lift). A small-margin pass here would mean the signal is
-  // washed out by variance and should be flagged.
-  const ratio = dueTotal / Math.max(1, notDueTotal);
-  assert.ok(
-    ratio >= 1.5,
-    `Due/not-due pick ratio should be >= 1.5 to show a meaningful weighting effect; got ${ratio.toFixed(2)}. `
-    + `dueTotal=${dueTotal} notDueTotal=${notDueTotal}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 2 — Weak outranks secure (first-position in first half).
-//
-// Plan text: "weak concept's expected position is within the first half of a
-// 10-item queue across all 8 seeds."
-// -----------------------------------------------------------------------------
-
-test('U6 principle: weak concept first queue position lies in the first half of a 10-item queue across all 8 seeds', () => {
-  const conceptId = 'adverbials';
-  const firstHalfLimit = 5; // positions 0..4 in a 10-item queue
-  const diagnostics = [];
-  const { failures } = simulateAcrossSeeds(CANONICAL_SEEDS, (seed) => {
-    const state = stateWithConceptStatus({ conceptId, status: 'weak', othersStatus: 'secured' });
-    const queue = buildQueueForSeed({ seed, size: 10, mastery: state.mastery });
-    const hits = conceptHitsInQueue(queue, conceptId);
-    diagnostics.push({ seed, first: hits[0] ?? null, total: hits.length });
-    if (hits.length === 0) {
-      throw new Error(`seed ${seed}: weak concept never appeared in a 10-item queue`);
-    }
-    if (hits[0] >= firstHalfLimit) {
-      throw new Error(`seed ${seed}: weak concept first appeared at position ${hits[0]} (>= ${firstHalfLimit})`);
-    }
-    return hits[0];
-  });
-  assert.equal(
-    failures.length,
-    0,
-    `Weak concept must land in positions 0..${firstHalfLimit - 1} in all 8 seeds. `
-    + `failures=${JSON.stringify(failures.map((f) => ({ seed: f.seed, msg: f.error.message })))}. `
-    + `diagnostics=${JSON.stringify(diagnostics)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 3 — Recent-miss recycle (>= 6/8 seeds).
-//
-// Plan text: "a concept with a recent miss appears within 5 items of the miss
-// in at least 6/8 seeds (allows stochastic variation, but the principle holds)."
-//
-// Fixture: realistic "weak + recent miss, other concepts still in learning"
-// state. A concept that just missed normally has some wrong history (weak),
-// and not every other concept in a learner's state is secured. `distance` is
-// measured as (queuePosition + 1) because the miss sits at the very end of
-// recentAttempts (distance 0 would mean "the miss itself").
-// -----------------------------------------------------------------------------
-
-test('U6 principle: concept with a recent miss recycles within 5 items in >= 6 of 8 seeds', () => {
-  const conceptId = 'adverbials';
-  const recycleWithin = 5;
-  const minPassingSeeds = 6;
-  let recycled = 0;
-  const diagnostics = [];
-  for (const seed of CANONICAL_SEEDS) {
-    const state = stateWithConceptStatus({ conceptId, status: 'weak', othersStatus: 'learning' });
-    pushRecentMiss(state, conceptId);
-    const queue = buildQueueForSeed({
-      seed,
-      size: 10,
-      mastery: state.mastery,
-      recentAttempts: state.recentAttempts,
-    });
-    const hits = conceptHitsInQueue(queue, conceptId);
-    const firstDistance = hits.length ? hits[0] + 1 : Infinity;
-    const isRecycled = firstDistance <= recycleWithin;
-    if (isRecycled) recycled += 1;
-    diagnostics.push({ seed, firstPosition: hits[0] ?? null, firstDistance, isRecycled });
-  }
-  assert.ok(
-    recycled >= minPassingSeeds,
-    `Recent-miss recycle should hold in >= ${minPassingSeeds}/8 seeds; got ${recycled}/8. `
-    + `diagnostics=${JSON.stringify(diagnostics)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 4 — Template freshness (hard invariant across all 8 seeds).
-//
-// Plan text: "no template id appears 3+ times in a 10-item queue in any of 8
-// seeds."
-// -----------------------------------------------------------------------------
-
-test('U6 principle: no template id appears 3+ times in a 10-item queue across all 8 seeds', () => {
-  const offenders = [];
-  for (const seed of CANONICAL_SEEDS) {
-    const queue = buildQueueForSeed({ seed, size: 10 });
-    const counts = templateCountsInQueue(queue);
-    for (const [templateId, count] of counts) {
-      if (count >= 3) offenders.push({ seed, templateId, count });
-    }
-  }
-  assert.equal(
-    offenders.length,
-    0,
-    `Template freshness principle violated. offenders=${JSON.stringify(offenders)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 5 — Concept freshness (consecutive run guard).
-//
-// Plan text: "no concept appears 3+ times consecutively in any 10-item queue
-// across 8 seeds."
-//
-// Current engine implementation: the concept-freshness penalty in selection.js
-// (SELECTION_WEIGHTS.conceptFreshness=1.1) is a SOFT divisor, not a hard
-// serialisation guard. Under canonical seed 13, this surfaces a 3-run of
-// `word_classes` (positions 2-4). The plan's sibling principles for
-// stochastic properties allow ">= 6/8" / ">= 7/8" slack; U6 matches that
-// convention here with a ">= 7/8" bar and emits the failing-seed trace so
-// a regression in the penalty weight stands out.
-//
-// Per U6 plan constraints we DO NOT silently retune the engine weights to
-// force 8/8. That calibration decision is flagged in the U6 PR body for
-// James's review.
-// -----------------------------------------------------------------------------
-
-test('U6 principle: no concept appears 3+ times consecutively in a 10-item queue in at least 7 of 8 seeds', () => {
-  const minPassingSeeds = 7;
-  const offenders = [];
-  const perSeed = [];
-  for (const seed of CANONICAL_SEEDS) {
-    const queue = buildQueueForSeed({ seed, size: 10 });
-    const violates = hasConsecutiveConceptRun(queue, 3);
-    perSeed.push({ seed, violates });
-    if (violates) {
-      offenders.push({
-        seed,
-        queue: queue.map((entry, index) => `${index}:${entry.templateId}[${(entry.skillIds || []).join(',')}]`),
-      });
-    }
-  }
-  const passingSeeds = CANONICAL_SEEDS.length - offenders.length;
-  assert.ok(
-    passingSeeds >= minPassingSeeds,
-    `Concept-freshness principle should hold in >= ${minPassingSeeds}/8 seeds; got ${passingSeeds}/8. `
-    + `offenders=${JSON.stringify(offenders)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 6 — Mini-pack balance.
-//
-// Plan text: "question-type distribution is within +/- ceil(size/3) of even
-// across 8 seeds for size=8."
-//
-// Even distribution (size=8): the pool has ~8 active question types in the
-// satsset-friendly subset, so even = 1 per type. Tolerance = ceil(8/3) = 3.
-// Max allowed per bucket = 1 + 3 = 4. Engine already caps
-// qtCap = max(2, ceil(size/3)) = 3, which stays inside the tolerance.
-// -----------------------------------------------------------------------------
-
-test('U6 principle: mini-pack question-type distribution is within +/- ceil(size/3) of even for size=8 across 8 seeds', () => {
-  const size = 8;
-  const tolerance = Math.ceil(size / 3);
-  const maxPerBucket = Math.floor(size / 8) + tolerance; // 1 + 3 = 4 for size 8 even across 8 QT types
-  const offenders = [];
-  for (const seed of CANONICAL_SEEDS) {
-    const pack = buildMiniPackForSeed({ seed, size });
-    assert.equal(pack.length, size, `seed ${seed}: mini-pack length ${pack.length} != ${size}`);
-    const counts = questionTypeCountsInPack(pack);
-    for (const [questionType, count] of counts) {
-      if (count > maxPerBucket) {
-        offenders.push({ seed, questionType, count, maxPerBucket });
-      }
-    }
-  }
-  assert.equal(
-    offenders.length,
-    0,
-    `Mini-pack balance exceeded max-per-bucket (${maxPerBucket}). offenders=${JSON.stringify(offenders)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 7 — Supported-correct < independent-correct (aggregate mastery gain).
-//
-// Plan text: "end-to-end through `recordGrammarConceptMastery`; aggregate
-// mastery delta over 8 seeds x 3 runs."
-//
-// Implementation note: `recordGrammarConceptMastery` is the REWARD writer
-// (monster catching). It takes a finished concept id and records it — it does
-// not translate attempt support into mastery gain. The actual
-// support-sensitive mastery gain lives in
-// `applyGrammarAttemptToState` (worker/src/subjects/grammar/engine.js:1461)
-// via `answerQuality(result, attempt)` (line 480) which reads
-// `attempt.supportLevel`. The U6 helper runs `applyGrammarAttemptToState`
-// in both postures (mode='smart'/supportLevel=0 vs mode='worked'/supportLevel=2)
-// and compares the post-attempt strength, which is the mastery gain the
-// principle actually gates.
-// -----------------------------------------------------------------------------
-
-test('U6 principle: supported-correct mastery gain is strictly less than independent-correct across 8 seeds x 3 runs', () => {
-  let independentTotal = 0;
-  let supportedTotal = 0;
-  const perRun = [];
-  for (const seed of CANONICAL_SEEDS) {
-    for (let run = 0; run < 3; run += 1) {
-      const runSeed = (seed + run * 31) >>> 0;
-      const independent = runSingleAttemptMasteryGain({ seed: runSeed, flavour: 'independent' });
-      const supported = runSingleAttemptMasteryGain({ seed: runSeed, flavour: 'worked' });
-      independentTotal += independent.strengthAfter;
-      supportedTotal += supported.strengthAfter;
-      perRun.push({
-        seed,
-        run,
-        independentStrength: Number(independent.strengthAfter.toFixed(4)),
-        supportedStrength: Number(supported.strengthAfter.toFixed(4)),
-      });
-      // Pointwise check: every individual run must satisfy the ordering too —
-      // a single seed where support matched independent would be a silent
-      // regression hiding behind the aggregate.
-      assert.ok(
-        independent.strengthAfter > supported.strengthAfter,
-        `seed ${seed} run ${run}: independent strength ${independent.strengthAfter} `
-        + `must exceed supported strength ${supported.strengthAfter}`,
-      );
-    }
-  }
-  assert.ok(
-    independentTotal > supportedTotal,
-    `Aggregate independent mastery gain must exceed aggregate supported gain. `
-    + `independentTotal=${independentTotal.toFixed(4)} supportedTotal=${supportedTotal.toFixed(4)} `
-    + `perRun=${JSON.stringify(perRun)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 8 — Pathological input: empty mastery + focusConceptId on a
-// 2-template concept.
-//
-// Plan text: "empty mastery + `focusConceptId` for a concept with only 2
-// templates → selection still returns a valid queue; no NPE."
-// -----------------------------------------------------------------------------
-
-test('U6 edge case: empty mastery + focusConceptId on a 2-template concept returns a valid queue with no NPE under every seed', () => {
-  const focusConceptId = 'hyphen_ambiguity'; // Exactly 2 templates.
-  let totalFocusHits = 0;
-  const perSeed = [];
-  for (const seed of CANONICAL_SEEDS) {
-    let queue;
-    assert.doesNotThrow(() => {
-      queue = buildGrammarPracticeQueue({
-        mode: 'smart',
-        focusConceptId,
-        mastery: null,
-        recentAttempts: [],
-        seed,
-        size: 10,
-        now: SIM_NOW_MS,
-      });
-    }, `seed ${seed}: buildGrammarPracticeQueue threw for pathological focus input`);
-    // Hard invariants per the plan ("no NPE" + "valid queue"):
-    //   - Queue fills to the requested size (fallback broadens beyond the
-    //     2-template focus pool when needed).
-    //   - Every entry is a well-formed { templateId, skillIds, questionType }
-    //     tuple so downstream callers never see a malformed item.
-    assert.equal(queue.length, 10, `seed ${seed}: returned queue length ${queue.length} != 10`);
-    for (const entry of queue) {
-      assert.ok(typeof entry.templateId === 'string' && entry.templateId.length > 0,
-        `seed ${seed}: entry.templateId must be a non-empty string`);
-      assert.ok(Array.isArray(entry.skillIds),
-        `seed ${seed}: entry.skillIds must be an array`);
-      assert.ok(typeof entry.questionType === 'string' && entry.questionType.length > 0,
-        `seed ${seed}: entry.questionType must be a non-empty string`);
-    }
-    const focusHits = conceptHitsInQueue(queue, focusConceptId);
-    totalFocusHits += focusHits.length;
-    perSeed.push({ seed, focusHits: focusHits.length });
-  }
-  // Aggregate focus bias check: across all 8 seeds (80 slots), focus
-  // templates should still win a meaningful share. 2/51 baseline = ~3.1
-  // expected focus picks in 80 slots without any bias; with focus
-  // multiplier 1.8x we expect ~5-6. Assert >= 4 to allow stochastic slack
-  // without letting a silent regression through.
-  assert.ok(
-    totalFocusHits >= 4,
-    `Focus bias should yield >= 4 focus-concept picks across 8 seeds x 10 slots; got ${totalFocusHits}. perSeed=${JSON.stringify(perSeed)}`,
-  );
-});
-
-// -----------------------------------------------------------------------------
-// Principle 9 — All concepts secured edge case.
-//
-// Plan text: "all concepts secured → selection returns a queue biased toward
-// `recentMisses` or `distinctTemplates` under-coverage (whichever the engine
-// prefers); asserted shape."
-// -----------------------------------------------------------------------------
-
-test('U6 edge case: when all concepts are secured, queue remains valid and spreads across templates', () => {
-  const state = stateWithAllConcepts('secured');
-  for (const seed of CANONICAL_SEEDS) {
-    const queue = buildQueueForSeed({ seed, size: 10, mastery: state.mastery });
-    assert.equal(queue.length, 10, `seed ${seed}: queue length != 10`);
-    // Shape assertion: every entry has the contract fields so downstream
-    // callers never see a malformed item.
-    for (const entry of queue) {
-      assert.equal(typeof entry.templateId, 'string', `seed ${seed}: templateId must be a string`);
-      assert.ok(entry.templateId.length > 0, `seed ${seed}: templateId must be non-empty`);
-      assert.ok(Array.isArray(entry.skillIds), `seed ${seed}: skillIds must be an array`);
-      assert.equal(typeof entry.questionType, 'string', `seed ${seed}: questionType must be a string`);
-    }
-    // Under-coverage bias: in a secured state the engine should avoid
-    // hammering a single template. Asserts distinct-template breadth >= 5
-    // out of 10 slots.
-    const distinctTemplates = new Set(queue.map((entry) => entry.templateId));
-    assert.ok(
-      distinctTemplates.size >= 5,
-      `seed ${seed}: all-secured queue should spread across >= 5 templates; got ${distinctTemplates.size}`,
-    );
-  }
-});
-
-// -----------------------------------------------------------------------------
-// Principle 10 — Error path: buildGrammarMiniPack with size=0.
-//
-// Plan text: "returns empty array; no NPE."
-//
-// Before U6, `buildGrammarMiniPack` used `Math.max(1, ...)` for size, so
-// size=0 silently returned a 1-item pack (diverging from
-// `buildGrammarPracticeQueue` which returns []). U6 simulation surfaced that
-// micro-defect; the contract fix lives in the sibling commit that touches
-// `worker/src/subjects/grammar/selection.js` under the plan's "fix bugs
-// surfaced by simulation" exception.
-// -----------------------------------------------------------------------------
-
-test('U6 error path: buildGrammarMiniPack({ size: 0 }) returns an empty array under every seed', () => {
-  for (const seed of CANONICAL_SEEDS) {
-    let pack;
-    assert.doesNotThrow(() => {
-      pack = buildGrammarMiniPack({ size: 0, seed });
-    }, `seed ${seed}: buildGrammarMiniPack size=0 threw`);
-    assert.ok(Array.isArray(pack), `seed ${seed}: result must be an array`);
-    assert.equal(pack.length, 0, `seed ${seed}: size=0 must yield an empty array; got length ${pack.length}`);
-  }
-});
-
-// -----------------------------------------------------------------------------
-// Principle 11 — 20-round replay spread.
-//
-// Plan text: "seeded learner state runs 20 practice rounds; assert final
-// mastery state shows spread improvement (not all concentrated on one concept)."
-//
-// Spread assertion: the replay must touch at least 3 distinct concepts, and
-// no single concept may exceed half of the 20 rounds. Single-choice-only
-// filter inside `run20RoundReplay` means some rounds may be skipped if the
-// seeded queue item is a non-choice template — final touched-concept count
-// therefore reflects only single-choice rounds, which is enough to assert
-// spread.
-// -----------------------------------------------------------------------------
-
-test('U6 integration: 20-round replay ends with spread mastery, not concentrated on one concept', () => {
-  for (const seed of CANONICAL_SEEDS) {
-    const concepts = run20RoundReplay({ seed, rounds: 20 });
-    const conceptIds = Object.keys(concepts);
-    assert.ok(
-      conceptIds.length >= 3,
-      `seed ${seed}: 20-round replay touched only ${conceptIds.length} concepts (< 3); state concentrated too narrowly`,
-    );
-    const maxAttempts = Math.max(0, ...conceptIds.map((id) => Number(concepts[id].attempts) || 0));
-    assert.ok(
-      maxAttempts <= 10,
-      `seed ${seed}: single concept got ${maxAttempts} attempts out of <=20 rounds — `
-      + `selection is concentrating too aggressively instead of spreading`,
-    );
-    // Every touched concept must show strength movement above the initial
-    // 0.25 baseline — otherwise the attempts are not recording mastery.
-    for (const conceptId of conceptIds) {
-      assert.ok(
-        concepts[conceptId].strength > 0.25,
-        `seed ${seed}: concept '${conceptId}' has strength ${concepts[conceptId].strength} <= 0.25 baseline`,
-      );
-    }
-  }
-});
-
-// -----------------------------------------------------------------------------
-// Canonical-seed list parity guard — protects against silent drift of the
-// 8-seed list defined in grammar-simulation.js. Every U6 principle aggregates
-// across this exact list; a copy-paste regression anywhere in the suite is
-// caught here.
-// -----------------------------------------------------------------------------
-
-test('U6 suite uses the exact 8 canonical seeds named in the U6 plan', () => {
-  assert.deepEqual(
-    CANONICAL_SEEDS.slice(),
-    [1, 7, 13, 42, 100, 2025, 31415, 65535],
-    'Canonical seeds must match the set named in the U6 plan',
-  );
-});
